+++ package.js
@@ -3,12 +3,12 @@
 Package.describe({
   name: 'mizzao:user-status',
   summary: 'User connection and idle state tracking for Meteor',
-  version: '2.0.0-rc.2',
+  version: '2.0.0-rc.3',
   git: 'https://github.com/Meteor-Community-Packages/meteor-user-status.git'
 });
 
 Package.onUse((api) => {
-  api.versionsFrom(['2.8.1', '2.15', '3.0']);
+  api.versionsFrom(['2.16', '3.0']);
 
   api.use('ecmascript');
   api.use('accounts-base');
@@ -16,7 +16,7 @@ Package.onUse((api) => {
   api.use('mongo');
   api.use('ddp');
   api.use('tracker', 'client');
-  api.use('mizzao:timesync@0.5.5 || 1.0.0-beta.0');
+  api.use('mizzao:timesync@0.6.0 || 1.0.0-beta.2');
 
   api.export('MonitorInternals', 'client', {
     testOnly: true
@@ -33,11 +33,12 @@ Package.onUse((api) => {
 Package.onTest((api) => {
   api.use('ecmascript');
   api.use('mizzao:user-status');
-  api.use('mizzao:timesync@0.5.5 || 1.0.0-beta.0');
+  api.use('mizzao:timesync@0.6.0 || 1.0.0-beta.2');
 
   api.use(['accounts-base', 'accounts-password']);
 
   api.use(['random', 'tracker']);
+  api.use('jquery@3.0.2', 'client');
 
   api.use('test-helpers');
   api.use('tinytest');