+++ tests/status_tests.js
@@ -3,7 +3,7 @@ import { Meteor } from 'meteor/meteor';
 import { Random } from 'meteor/random';
 import { Tracker } from 'meteor/tracker';
 import { StatusInternals, UserStatus } from '../server/status';
-import { getCleanupWrapper, TEST_IP, TEST_userId } from './setup';
+import { ensureTestUser, getCleanupWrapper, TEST_IP, TEST_userId } from './setup';
 
 let lastLoginAdvice = null;
 let lastLogoutAdvice = null;
@@ -18,8 +18,9 @@ UserStatus.events.on('connectionActive', advice => lastActiveAdvice = advice);
 
 const TEST_UA = 'old-ass browser';
 
+let lastDateTime = Date.now();
 // Make sure repeated calls to this return different values
-const delayedDate = () => Meteor.wrapAsync(cb => Meteor.setTimeout((() => cb(undefined, new Date())), 1))();
+const delayedDate = () => new Date(++lastDateTime);
 
 const randomConnection = () => ({
   id: Random.id(),
@@ -32,18 +33,22 @@ const randomConnection = () => ({
 
 // Delete the entire status field and sessions after each test
 const withCleanup = getCleanupWrapper({
-  after() {
+  async before() {
+    await ensureTestUser();
+  },
+
+  async after() {
     lastLoginAdvice = null;
     lastLogoutAdvice = null;
     lastIdleAdvice = null;
     lastActiveAdvice = null;
 
-    Meteor.users.updateAsync(TEST_userId, {
+    await Meteor.users.updateAsync(TEST_userId, {
       $unset: {
         status: null
       }
     });
-    UserStatus.connections.removeAsync({
+    await UserStatus.connections.removeAsync({
       $or: [{
           userId: TEST_userId
         },
@@ -60,10 +65,10 @@ const withCleanup = getCleanupWrapper({
 // Clean up before we add any tests just in case some crap left over from before
 withCleanup(function () {});
 
-Tinytest.add('status - adding anonymous session', withCleanup(async (test) => {
+Tinytest.addAsync('status - adding anonymous session', withCleanup(async (test) => {
   const conn = randomConnection();
 
-  StatusInternals.addSession(conn);
+  await StatusInternals.addSession(conn);
 
   const doc = await UserStatus.connections.findOneAsync(conn.id);
 
@@ -75,21 +80,22 @@ Tinytest.add('status - adding anonymous session', withCleanup(async (test) => {
   return test.isFalse(doc.loginTime);
 }));
 
-Tinytest.add('status - adding and removing anonymous session', withCleanup((test) => {
+Tinytest.addAsync('status - adding and removing anonymous session', withCleanup(async (test) => {
   const conn = randomConnection();
 
-  StatusInternals.addSession(conn);
-  StatusInternals.removeSession(conn, delayedDate());
+  await StatusInternals.addSession(conn);
+  await StatusInternals.removeSession(conn, delayedDate());
 
-  return test.isFalse(UserStatus.connections.findOne(conn.id));
+  const doc = await UserStatus.connections.findOneAsync(conn.id);
+  return test.isFalse(doc != null);
 }));
 
-Tinytest.add('status - adding one authenticated session', withCleanup(async(test) => {
+Tinytest.addAsync('status - adding one authenticated session', withCleanup(async(test) => {
   const conn = randomConnection();
   const ts = delayedDate();
 
-  StatusInternals.addSession(conn);
-  StatusInternals.loginSession(conn, ts, TEST_userId);
+  await StatusInternals.addSession(conn);
+  await StatusInternals.loginSession(conn, ts, TEST_userId);
 
   const doc = await UserStatus.connections.findOneAsync(conn.id);
   const user = await Meteor.users.findOneAsync(TEST_userId);
@@ -113,18 +119,18 @@ Tinytest.add('status - adding one authenticated session', withCleanup(async(test
   return test.equal(user.status.lastLogin.userAgent, TEST_UA);
 }));
 
-Tinytest.add('status - adding and removing one authenticated session', withCleanup(async(test) => {
+Tinytest.addAsync('status - adding and removing one authenticated session', withCleanup(async(test) => {
   const conn = randomConnection();
   const ts = delayedDate();
 
-  StatusInternals.addSession(conn);
-  StatusInternals.loginSession(conn, ts, TEST_userId);
+  await StatusInternals.addSession(conn);
+  await StatusInternals.loginSession(conn, ts, TEST_userId);
 
   const logoutTime = delayedDate();
-  StatusInternals.removeSession(conn, logoutTime);
+  await StatusInternals.removeSession(conn, logoutTime);
 
-  const doc = UserStatus.connections.findOneAsync(conn.id);
-  const user = Meteor.users.findOneAsync(TEST_userId);
+  const doc = await UserStatus.connections.findOneAsync(conn.id);
+  const user = await Meteor.users.findOneAsync(TEST_userId);
 
   test.isFalse(doc != null);
 
@@ -138,15 +144,15 @@ Tinytest.add('status - adding and removing one authenticated session', withClean
   return test.equal(user.status.lastLogin.date, ts);
 }));
 
-Tinytest.add('status - logout and then close one authenticated session', withCleanup((test) => {
+Tinytest.addAsync('status - logout and then close one authenticated session', withCleanup(async (test) => {
   const conn = randomConnection();
   const ts = delayedDate();
 
-  StatusInternals.addSession(conn);
-  StatusInternals.loginSession(conn, ts, TEST_userId);
+  await StatusInternals.addSession(conn);
+  await StatusInternals.loginSession(conn, ts, TEST_userId);
 
   const logoutTime = delayedDate();
-  StatusInternals.tryLogoutSession(conn, logoutTime);
+  await StatusInternals.tryLogoutSession(conn, logoutTime);
 
   test.equal(lastLogoutAdvice.userId, TEST_userId);
   test.equal(lastLogoutAdvice.connectionId, conn.id);
@@ -157,10 +163,10 @@ Tinytest.add('status - logout and then close one authenticated session', withCle
   // After logging out, the user closes the browser, which triggers a close callback
   // However, the event should not be emitted again
   const closeTime = delayedDate();
-  StatusInternals.removeSession(conn, closeTime);
+  await StatusInternals.removeSession(conn, closeTime);
 
-  const doc = UserStatus.connections.findOne(conn.id);
-  const user = Meteor.users.findOne(TEST_userId);
+  const doc = await UserStatus.connections.findOneAsync(conn.id);
+  const user = await Meteor.users.findOneAsync(TEST_userId);
 
   test.isFalse(doc != null);
   test.isFalse(lastLogoutAdvice != null);
@@ -170,16 +176,16 @@ Tinytest.add('status - logout and then close one authenticated session', withCle
   return test.equal(user.status.lastLogin.date, ts);
 }));
 
-Tinytest.add('status - idling one authenticated session', withCleanup(async(test) => {
+Tinytest.addAsync('status - idling one authenticated session', withCleanup(async(test) => {
   const conn = randomConnection();
   const ts = delayedDate();
 
-  StatusInternals.addSession(conn);
-  StatusInternals.loginSession(conn, ts, TEST_userId);
+  await StatusInternals.addSession(conn);
+  await StatusInternals.loginSession(conn, ts, TEST_userId);
 
   const idleTime = delayedDate();
 
-  StatusInternals.idleSession(conn, idleTime, TEST_userId);
+  await StatusInternals.idleSession(conn, idleTime, TEST_userId);
 
   const doc = await UserStatus.connections.findOneAsync(conn.id);
   const user = await Meteor.users.findOneAsync(TEST_userId);
@@ -203,17 +209,17 @@ Tinytest.add('status - idling one authenticated session', withCleanup(async(test
   return test.equal(user.status.lastActivity, idleTime);
 }));
 
-Tinytest.add('status - idling and reactivating one authenticated session', withCleanup(async(test) => {
+Tinytest.addAsync('status - idling and reactivating one authenticated session', withCleanup(async(test) => {
   const conn = randomConnection();
   const ts = delayedDate();
 
-  StatusInternals.addSession(conn);
-  StatusInternals.loginSession(conn, ts, TEST_userId);
+  await StatusInternals.addSession(conn);
+  await StatusInternals.loginSession(conn, ts, TEST_userId);
 
   const idleTime = delayedDate();
-  StatusInternals.idleSession(conn, idleTime, TEST_userId);
+  await StatusInternals.idleSession(conn, idleTime, TEST_userId);
   const activeTime = delayedDate();
-  StatusInternals.activeSession(conn, activeTime, TEST_userId);
+  await StatusInternals.activeSession(conn, activeTime, TEST_userId);
 
   const doc = await UserStatus.connections.findOneAsync(conn.id);
   const user = await Meteor.users.findOneAsync(TEST_userId);
@@ -237,16 +243,16 @@ Tinytest.add('status - idling and reactivating one authenticated session', withC
   return test.isFalse(user.status.lastActivity != null);
 }));
 
-Tinytest.add('status - idling and removing one authenticated session', withCleanup(async(test) => {
+Tinytest.addAsync('status - idling and removing one authenticated session', withCleanup(async(test) => {
   const conn = randomConnection();
   const ts = delayedDate();
 
-  StatusInternals.addSession(conn);
-  StatusInternals.loginSession(conn, ts, TEST_userId);
+  await StatusInternals.addSession(conn);
+  await StatusInternals.loginSession(conn, ts, TEST_userId);
   const idleTime = delayedDate();
-  StatusInternals.idleSession(conn, idleTime, TEST_userId);
+  await StatusInternals.idleSession(conn, idleTime, TEST_userId);
   const logoutTime = delayedDate();
-  StatusInternals.removeSession(conn, logoutTime);
+  await StatusInternals.removeSession(conn, logoutTime);
 
   const doc = await UserStatus.connections.findOneAsync(conn.id);
   const user = await Meteor.users.findOneAsync(TEST_userId);
@@ -262,26 +268,26 @@ Tinytest.add('status - idling and removing one authenticated session', withClean
   return test.equal(user.status.lastLogin.date, ts);
 }));
 
-Tinytest.add('status - idling and reconnecting one authenticated session', withCleanup(async(test) => {
+Tinytest.addAsync('status - idling and reconnecting one authenticated session', withCleanup(async(test) => {
   const conn = randomConnection();
   const ts = delayedDate();
 
-  StatusInternals.addSession(conn);
-  StatusInternals.loginSession(conn, ts, TEST_userId);
+  await StatusInternals.addSession(conn);
+  await StatusInternals.loginSession(conn, ts, TEST_userId);
   const idleTime = delayedDate();
-  StatusInternals.idleSession(conn, idleTime, TEST_userId);
+  await StatusInternals.idleSession(conn, idleTime, TEST_userId);
 
   // Session reconnects but was idle
 
   const discTime = delayedDate();
-  StatusInternals.removeSession(conn, discTime);
+  await StatusInternals.removeSession(conn, discTime);
 
   const reconn = randomConnection();
   const reconnTime = delayedDate();
 
-  StatusInternals.addSession(reconn);
-  StatusInternals.loginSession(reconn, reconnTime, TEST_userId);
-  StatusInternals.idleSession(reconn, idleTime, TEST_userId);
+  await StatusInternals.addSession(reconn);
+  await StatusInternals.loginSession(reconn, reconnTime, TEST_userId);
+  await StatusInternals.idleSession(reconn, idleTime, TEST_userId);
 
   const doc = await UserStatus.connections.findOneAsync(reconn.id);
   const user = await Meteor.users.findOneAsync(TEST_userId);
@@ -301,7 +307,7 @@ Tinytest.add('status - idling and reconnecting one authenticated session', withC
   return test.equal(user.status.lastActivity, idleTime);
 }));
 
-Tinytest.add('multiplex - two online sessions', withCleanup(async (test) => {
+Tinytest.addAsync('multiplex - two online sessions', withCleanup(async (test) => {
   const conn = randomConnection();
 
   const conn2 = randomConnection();
@@ -309,11 +315,11 @@ Tinytest.add('multiplex - two online sessions', withCleanup(async (test) => {
   const ts = delayedDate();
   const ts2 = delayedDate();
 
-  StatusInternals.addSession(conn);
-  StatusInternals.loginSession(conn, ts, TEST_userId);
+  await StatusInternals.addSession(conn);
+  await StatusInternals.loginSession(conn, ts, TEST_userId);
 
-  StatusInternals.addSession(conn2);
-  StatusInternals.loginSession(conn2, ts2, TEST_userId);
+  await StatusInternals.addSession(conn2);
+  await StatusInternals.loginSession(conn2, ts2, TEST_userId);
 
   const user = await Meteor.users.findOneAsync(TEST_userId);
 
@@ -321,8 +327,7 @@ Tinytest.add('multiplex - two online sessions', withCleanup(async (test) => {
   return test.equal(user.status.lastLogin.date, ts2);
 }));
 
-Tinytest.add('multiplex - two online sessions with one going offline', withCleanup(async (test) => {
-  let user;
+Tinytest.addAsync('multiplex - two online sessions with one going offline', withCleanup(async (test) => {
   const conn = randomConnection();
 
   const conn2 = randomConnection();
@@ -330,21 +335,20 @@ Tinytest.add('multiplex - two online sessions with one going offline', withClean
   const ts = delayedDate();
   const ts2 = delayedDate();
 
-  StatusInternals.addSession(conn);
-  StatusInternals.loginSession(conn, ts, TEST_userId);
+  await StatusInternals.addSession(conn);
+  await StatusInternals.loginSession(conn, ts, TEST_userId);
 
-  StatusInternals.addSession(conn2);
-  StatusInternals.loginSession(conn2, ts2, TEST_userId);
+  await StatusInternals.addSession(conn2);
+  await StatusInternals.loginSession(conn2, ts2, TEST_userId);
 
-  StatusInternals.removeSession(conn, delayedDate(),
-
-    (user = await Meteor.users.findOneAsync(TEST_userId)));
+  await StatusInternals.removeSession(conn, delayedDate());
+  const user = await Meteor.users.findOneAsync(TEST_userId);
 
   test.equal(user.status.online, true);
   return test.equal(user.status.lastLogin.date, ts2);
 }));
 
-Tinytest.add('multiplex - two online sessions to offline', withCleanup(async (test) => {
+Tinytest.addAsync('multiplex - two online sessions to offline', withCleanup(async (test) => {
   const conn = randomConnection();
 
   const conn2 = randomConnection();
@@ -352,14 +356,14 @@ Tinytest.add('multiplex - two online sessions to offline', withCleanup(async (te
   const ts = delayedDate();
   const ts2 = delayedDate();
 
-  StatusInternals.addSession(conn);
-  StatusInternals.loginSession(conn, ts, TEST_userId);
+  await StatusInternals.addSession(conn);
+  await StatusInternals.loginSession(conn, ts, TEST_userId);
 
-  StatusInternals.addSession(conn2);
-  StatusInternals.loginSession(conn2, ts2, TEST_userId);
+  await StatusInternals.addSession(conn2);
+  await StatusInternals.loginSession(conn2, ts2, TEST_userId);
 
-  StatusInternals.removeSession(conn, delayedDate());
-  StatusInternals.removeSession(conn2, delayedDate());
+  await StatusInternals.removeSession(conn, delayedDate());
+  await StatusInternals.removeSession(conn2, delayedDate());
 
   const user = await Meteor.users.findOneAsync(TEST_userId);
 
@@ -367,7 +371,7 @@ Tinytest.add('multiplex - two online sessions to offline', withCleanup(async (te
   return test.equal(user.status.lastLogin.date, ts2);
 }));
 
-Tinytest.add('multiplex - idling one of two online sessions', withCleanup(async (test) => {
+Tinytest.addAsync('multiplex - idling one of two online sessions', withCleanup(async (test) => {
   const conn = randomConnection();
 
   const conn2 = randomConnection();
@@ -375,14 +379,14 @@ Tinytest.add('multiplex - idling one of two online sessions', withCleanup(async 
   const ts = delayedDate();
   const ts2 = delayedDate();
 
-  StatusInternals.addSession(conn);
-  StatusInternals.loginSession(conn, ts, TEST_userId);
+  await StatusInternals.addSession(conn);
+  await StatusInternals.loginSession(conn, ts, TEST_userId);
 
-  StatusInternals.addSession(conn2);
-  StatusInternals.loginSession(conn2, ts2, TEST_userId);
+  await StatusInternals.addSession(conn2);
+  await StatusInternals.loginSession(conn2, ts2, TEST_userId);
 
   const idle1 = delayedDate();
-  StatusInternals.idleSession(conn, idle1, TEST_userId);
+  await StatusInternals.idleSession(conn, idle1, TEST_userId);
 
   const user = await Meteor.users.findOneAsync(TEST_userId);
 
@@ -391,7 +395,7 @@ Tinytest.add('multiplex - idling one of two online sessions', withCleanup(async 
   return test.equal(user.status.idle, false);
 }));
 
-Tinytest.add('multiplex - idling two online sessions', withCleanup(async (test) => {
+Tinytest.addAsync('multiplex - idling two online sessions', withCleanup(async (test) => {
   const conn = randomConnection();
 
   const conn2 = randomConnection();
@@ -399,16 +403,16 @@ Tinytest.add('multiplex - idling two online sessions', withCleanup(async (test) 
   const ts = delayedDate();
   const ts2 = delayedDate();
 
-  StatusInternals.addSession(conn);
-  StatusInternals.loginSession(conn, ts, TEST_userId);
+  await StatusInternals.addSession(conn);
+  await StatusInternals.loginSession(conn, ts, TEST_userId);
 
-  StatusInternals.addSession(conn2);
-  StatusInternals.loginSession(conn2, ts2, TEST_userId);
+  await StatusInternals.addSession(conn2);
+  await StatusInternals.loginSession(conn2, ts2, TEST_userId);
 
   const idle1 = delayedDate();
   const idle2 = delayedDate();
-  StatusInternals.idleSession(conn, idle1, TEST_userId);
-  StatusInternals.idleSession(conn2, idle2, TEST_userId);
+  await StatusInternals.idleSession(conn, idle1, TEST_userId);
+  await StatusInternals.idleSession(conn2, idle2, TEST_userId);
 
   const user = await Meteor.users.findOneAsync(TEST_userId);
 
@@ -418,7 +422,7 @@ Tinytest.add('multiplex - idling two online sessions', withCleanup(async (test) 
   return test.equal(user.status.lastActivity, idle2);
 }));
 
-Tinytest.add('multiplex - idling two then reactivating one session', withCleanup(async (test) => {
+Tinytest.addAsync('multiplex - idling two then reactivating one session', withCleanup(async (test) => {
   const conn = randomConnection();
 
   const conn2 = randomConnection();
@@ -426,18 +430,18 @@ Tinytest.add('multiplex - idling two then reactivating one session', withCleanup
   const ts = delayedDate();
   const ts2 = delayedDate();
 
-  StatusInternals.addSession(conn);
-  StatusInternals.loginSession(conn, ts, TEST_userId);
+  await StatusInternals.addSession(conn);
+  await StatusInternals.loginSession(conn, ts, TEST_userId);
 
-  StatusInternals.addSession(conn2);
-  StatusInternals.loginSession(conn2, ts2, TEST_userId);
+  await StatusInternals.addSession(conn2);
+  await StatusInternals.loginSession(conn2, ts2, TEST_userId);
 
   const idle1 = delayedDate();
   const idle2 = delayedDate();
-  StatusInternals.idleSession(conn, idle1, TEST_userId);
-  StatusInternals.idleSession(conn2, idle2, TEST_userId);
+  await StatusInternals.idleSession(conn, idle1, TEST_userId);
+  await StatusInternals.idleSession(conn2, idle2, TEST_userId);
 
-  StatusInternals.activeSession(conn, delayedDate(), TEST_userId);
+  await StatusInternals.activeSession(conn, delayedDate(), TEST_userId);
 
   const user = await Meteor.users.findOneAsync(TEST_userId);
 
@@ -447,7 +451,7 @@ Tinytest.add('multiplex - idling two then reactivating one session', withCleanup
   return test.isFalse(user.status.lastActivity != null);
 }));
 
-Tinytest.add('multiplex - logging in while an existing session is idle', withCleanup(async (test) => {
+Tinytest.addAsync('multiplex - logging in while an existing session is idle', withCleanup(async (test) => {
   const conn = randomConnection();
 
   const conn2 = randomConnection();
@@ -455,14 +459,14 @@ Tinytest.add('multiplex - logging in while an existing session is idle', withCle
   const ts = delayedDate();
   const ts2 = delayedDate();
 
-  StatusInternals.addSession(conn);
-  StatusInternals.loginSession(conn, ts, TEST_userId);
+  await StatusInternals.addSession(conn);
+  await StatusInternals.loginSession(conn, ts, TEST_userId);
 
   const idle1 = delayedDate();
-  StatusInternals.idleSession(conn, idle1, TEST_userId);
+  await StatusInternals.idleSession(conn, idle1, TEST_userId);
 
-  StatusInternals.addSession(conn2);
-  StatusInternals.loginSession(conn2, ts2, TEST_userId);
+  await StatusInternals.addSession(conn2);
+  await StatusInternals.loginSession(conn2, ts2, TEST_userId);
 
   const user = await Meteor.users.findOneAsync(TEST_userId);
 
@@ -472,7 +476,7 @@ Tinytest.add('multiplex - logging in while an existing session is idle', withCle
   return test.isFalse(user.status.lastActivity != null);
 }));
 
-Tinytest.add('multiplex - simulate tab switch', withCleanup(async (test) => {
+Tinytest.addAsync('multiplex - simulate tab switch', withCleanup(async (test) => {
   const conn = randomConnection();
 
   const conn2 = randomConnection();
@@ -481,20 +485,20 @@ Tinytest.add('multiplex - simulate tab switch', withCleanup(async (test) => {
   const ts2 = delayedDate();
 
   // open first tab then becomes idle
-  StatusInternals.addSession(conn);
-  StatusInternals.loginSession(conn, ts, TEST_userId);
+  await StatusInternals.addSession(conn);
+  await StatusInternals.loginSession(conn, ts, TEST_userId);
 
   const idle1 = delayedDate();
-  StatusInternals.idleSession(conn, idle1, TEST_userId);
+  await StatusInternals.idleSession(conn, idle1, TEST_userId);
 
   // open second tab then becomes idle
-  StatusInternals.addSession(conn2);
-  StatusInternals.loginSession(conn2, ts2, TEST_userId);
+  await StatusInternals.addSession(conn2);
+  await StatusInternals.loginSession(conn2, ts2, TEST_userId);
   const idle2 = delayedDate();
-  StatusInternals.idleSession(conn2, idle2, TEST_userId);
+  await StatusInternals.idleSession(conn2, idle2, TEST_userId);
 
   // go back to first tab
-  StatusInternals.activeSession(conn, delayedDate(), TEST_userId);
+  await StatusInternals.activeSession(conn, delayedDate(), TEST_userId);
 
   const user = await Meteor.users.findOneAsync(TEST_userId);
 
@@ -505,7 +509,7 @@ Tinytest.add('multiplex - simulate tab switch', withCleanup(async (test) => {
 }));
 
 // Test for idling one session across a disconnection; not most recent idle time
-Tinytest.add('multiplex - disconnection and reconnection while idle', withCleanup(async (test) => {
+Tinytest.addAsync('multiplex - disconnection and reconnection while idle', withCleanup(async (test) => {
   const conn = randomConnection();
 
   const conn2 = randomConnection();
@@ -513,19 +517,19 @@ Tinytest.add('multiplex - disconnection and reconnection while idle', withCleanu
   const ts = delayedDate();
   const ts2 = delayedDate();
 
-  StatusInternals.addSession(conn);
-  StatusInternals.loginSession(conn, ts, TEST_userId);
+  await StatusInternals.addSession(conn);
+  await StatusInternals.loginSession(conn, ts, TEST_userId);
 
-  StatusInternals.addSession(conn2);
-  StatusInternals.loginSession(conn2, ts2, TEST_userId);
+  await StatusInternals.addSession(conn2);
+  await StatusInternals.loginSession(conn2, ts2, TEST_userId);
 
   const idle1 = delayedDate();
-  StatusInternals.idleSession(conn, idle1, TEST_userId);
+  await StatusInternals.idleSession(conn, idle1, TEST_userId);
   const idle2 = delayedDate();
-  StatusInternals.idleSession(conn2, idle2, TEST_userId);
+  await StatusInternals.idleSession(conn2, idle2, TEST_userId);
 
   // Second session, which connected later, reconnects but remains idle
-  StatusInternals.removeSession(conn2, delayedDate(), TEST_userId);
+  await StatusInternals.removeSession(conn2, delayedDate(), TEST_userId);
 
   let user = await Meteor.users.findOneAsync(TEST_userId);
 
@@ -537,10 +541,10 @@ Tinytest.add('multiplex - disconnection and reconnection while idle', withCleanu
   const reconn2 = randomConnection();
 
   const ts3 = delayedDate();
-  StatusInternals.addSession(reconn2);
-  StatusInternals.loginSession(reconn2, ts3, TEST_userId);
+  await StatusInternals.addSession(reconn2);
+  await StatusInternals.loginSession(reconn2, ts3, TEST_userId);
 
-  StatusInternals.idleSession(reconn2, idle2, TEST_userId);
+  await StatusInternals.idleSession(reconn2, idle2, TEST_userId);
 
   user = await Meteor.users.findOneAsync(TEST_userId);
 
@@ -550,9 +554,9 @@ Tinytest.add('multiplex - disconnection and reconnection while idle', withCleanu
   return test.equal(user.status.lastActivity, idle2);
 }));
 
-Tinytest.add('status - user online set to false on startup', withCleanup(async (test) => {
+Tinytest.addAsync('status - user online set to false on startup', withCleanup(async (test) => {
   // special argument to onStartup prevents this from affecting client tests
-  StatusInternals.onStartup(TEST_userId);
+  await StatusInternals.onStartup(TEST_userId);
 
   const userAfterStartup = await Meteor.users.findOneAsync(TEST_userId);
   // Check reset status
@@ -564,15 +568,15 @@ Tinytest.add('status - user online set to false on startup', withCleanup(async (
   const conn = randomConnection();
   const ts = delayedDate();
 
-  StatusInternals.addSession(conn);
-  StatusInternals.loginSession(conn, ts, TEST_userId);
+  await StatusInternals.addSession(conn);
+  await StatusInternals.loginSession(conn, ts, TEST_userId);
 
   const userAfterLogin = await Meteor.users.findOneAsync(TEST_userId);
 
   test.equal(userAfterLogin.status.online, true);
   test.isFalse(userAfterLogin.status.idle);
 
-  StatusInternals.onStartup(TEST_userId);
+  await StatusInternals.onStartup(TEST_userId);
 
   const userAfterRestart = await Meteor.users.findOneAsync(TEST_userId);
   // Check reset status again